Frames/clima-app/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
